+++ lib/routes/create_invoice/qr_code_dialog.dart
@@ -151,7 +151,7 @@ class QrCodeDialogState extends State<QrCodeDialog> with SingleTickerProviderSta
                           const Padding(padding: EdgeInsets.only(top: 16.0)),
                           SizedBox(
                             width: MediaQuery.of(context).size.width,
-                            child: ExpiryAndFeeMessage(feesSat: widget.prepareReceiveResponse.feesSat),
+                            child: ExpiryAndFeeMessage(feesSat: widget.prepareReceiveResponse.feesSat.toInt()),
                           ),
                           const Padding(padding: EdgeInsets.only(top: 16.0)),
                         ],
